add boilerplate to auth test & rename import path in sagas
